backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.32.*
pydantic==2.*
pydantic-settings==2.*
email-validator==2.*
pymongo==4.*
python-multipart==0.0.*
passlib[bcrypt]==1.7.*
bcrypt==4.0.*
pyjwt==2.*
python-dotenv==1.*
python-dateutil==2.*
